completions: add macOS/BSD detection for more coreutils
Like we do for others.

Part of #12997

+++ share/completions/rmdir.fish
@@ -1,7 +1,13 @@
 #Completions for rmdir
 complete -x -c rmdir -a "(__fish_complete_directories (commandline -ct))"
-complete -c rmdir -l ignore-fail-on-non-empty -d "Ignore errors from non-empty directories"
-complete -c rmdir -s p -l parents -d "Remove each component of path"
-complete -c rmdir -s v -l verbose -d "Verbose mode"
-complete -c rmdir -l help -d "Display help and exit"
-complete -c rmdir -l version -d "Display version and exit"
+set -l is_gnu
+__fish_supports_version rmdir; and set is_gnu --is-gnu
+
+__fish_gnu_complete -c rmdir -s p -l parents -d "Remove each component of path" $is_gnu
+__fish_gnu_complete -c rmdir -s v -l verbose -d "Verbose mode" $is_gnu
+
+if test -n "$is_gnu"
+    complete -c rmdir -l ignore-fail-on-non-empty -d "Ignore errors from non-empty directories"
+    complete -c rmdir -l help -d "Display help and exit"
+    complete -c rmdir -l version -d "Display version and exit"
+end
